int in write_full (off)

+++ user_wr.c
@@ -14,7 +14,7 @@ enum
 // Devuelve bytes escritos (>0 -> puede ser <len) <-> -1 en error
 static int write_full (int fd, const char *buf, int len)
 {
-  ssize_t off = 0;
+  int off = 0;
   int w = 0;
 
   while (off < len) {
